+++ src/main/java/racingcar/domain/Car.java
@@ -1,11 +1,11 @@
-package racingcar;
+package racingcar.domain;
 
 public class Car implements Comparable<Car> {
     private static final int INIT_POSITION = 1;
     private static final int LOWER_MOVABLE_BOUND = 4;
 
     private int position = INIT_POSITION;
-    private String name;
+    private final String name;
 
     public Car(String name) {
         this.name = name;
