+++ src/test/java/racingcar/CarTest.java
@@ -1,6 +1,7 @@
 package racingcar;
 
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
 
 import static org.assertj.core.api.Assertions.*;
