+++ src/test/java/racingcar/RacingGameTest.java
@@ -2,6 +2,8 @@ package racingcar;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.service.RacingGame;
 
 
 import java.util.ArrayList;
